requirements.txt
pydantic>=2.0.0
websockets>=11.0
aiohttp>=3.8.0
openai>=1.0.0
python-docx>=1.1.2
PyMuPDF>=1.24.9
SQLAlchemy>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv
python-multipart
# Database driver for PostgreSQL
psycopg2-binary 